requirements.txt
transformers[tf,torch,sentencepiece,vision,optuna,sklearn,onnxruntime]==4.11.3
seqeval==1.2.2
from-root==1.0.2
fastapi==0.78.0
accelerate==0.5.1
datasets==1.13.0
Pyaml